Disable fsck for btrfs as well.
The fsck.btrfs is not included by default, therefor fsck fails on first
boot. It's part of btrfs-tools (available for both wheezy and jessie), so if
that package is included, fsck.btrfs does work and /etc/fstab can be
configured accordingly.
But until that package is included by default, we'll disable fsck for
btrfs as well.