Fixed the issue of having an unbootable system when the root filesystem on USB.
This was mainly the cause when the root fstype was ext4, the default.
Hopefully this fixes issue #47.

This was caused by an too old kernel and/or the absense of /lib/modules
in the resulting image. By setting 'kernel_module' to true, the kernel
package will be installed and the various modules will be build.

This is not a complete fix, since the other filesystems don't
automatically work. If the fstype is btrfs, then adding btrfs-tools to
the packages option makes it work. With f2fs things are a bit trickier,
since there is a f2fs-tools package, but only for jessie, not wheezy.

But this will fix the default case (ext4) and in the meantime I'll work
on a better fix for the other filesystem types.