Added option to set roots public SSH key
A new option called rootssh may contain the public SSH key for the root
user, which will be placed in /root/.ssh/authorized_keys, so no
password is needed to login using SSH and the root user.
Providing this option also prevents changes to SSH config on jessie, so
root login using password will stay disabled on that distribution (as
it should be).