Added options to create a system user
Several new options have been added (create_user, userpw, userssh,
user_is_admin) that allow the script to create a new user on the
installed system. A password is set for the new user, optionally a
public SSH key is added to authorised_keys, and if user_is_admin is set
to yes/true, the user will be able to use sudo (which is added to the
installation) for root tasks.