Replaced boolean true with '1' to make it consistent with the rest.
Using '1' for usbroot and various ways to say true/false is inconsistent
and likely confusing. Since we can't change usbroot to use true/false,
since that would be an interface change, I've converted the others to
use '1' for true, just like usbroot.
Also fixed the check for usbroot itself, because you have to specify '1'
to enable it