rcS: add offset for boot/root partitions
The number is expressed in "blocks". That means a 512 byte unit.
Therefore aligning on 64MiB means "131072".

The partitions usually needs to be aligned, specially on SD cards. This
enables a basic alignment.

Fully automated alignement isn't done because the /boot patition is
usually of a fixed sized, and the offset should be the same for SD size.

Note that the default might be set to something like 131072, which
corresponds nicely to the default 50M /boot.