Verify the number of imported GPG keys
The `--import` command can be used to import keyrings in addition to
individual keys, so it is necessary to check that only the two desired
keys were imported. (Subsequent signature verifications depend on only
valid keys being in the default keyring.)

The current version is vulnerable to attackers who can MITM HTTPS
connections, because the attacker could replace `raspbian.public.key`
with a keyring containing both `raspbian.public.key` and the attacker's
key, then sign malicious files with their own key. The `update.sh`
script would check that `raspbian.public.key` was imported but not
notice the presence of the attacker's key. This commit prevents this
vulnerability by ensuring that only the two desired keys were imported.