Added 'libpam-systemd' to base_packages when the init system is systemd.
The 'libpam-systemd' package is a 'Recommends' for systemd and is
therefor normally installed on a Debian/Raspbian system. But
cdebootstrap doesn't automatically install recommends and without
libpam-systemd one can see problems with shutdown and when the user is
logged in through ssh, thus adding it to the base_packages list.