Ensure that DNS resolver can be used by rootfs tools
Now that resolvconf is in use, there is no resolver information in the rootfs
/etc/resolv.conf by default; this stops 'apt-get' from being able to download
package lists and packages when run via chroot by the installer script.

This patch copies the temporary resolv.conf from the installer into the rootfs;
it will be overwritten by resolvconf when the system is booted.