Add CPU revisions for model 3A+, 3B+ and 4B, remove CM ones.
Currently, the firmware+kernel from raspbian.org does NOT support the
RPi4, but detection should be present regardless of how that will
be dealt with.
Removed CPU revisions for Compute Module as we don't support it.
Also add RPi4 stanza to config.txt, even though it may need to be
changed.