Use 'echo', not 'return' to return a value and fix busybox tr call.
Busybox tr does not support character classes, so use a simpler form.
If you use 'return X', X needs to be a small integer and you need to
check its value with '$?', but if you use retval=$(function-call), then
you need to use 'echo' to return a string.
See https://stackoverflow.com/a/8743103 for details.