Replace cdebootstrap-static with its dynamic variant.
Primary argument to do this, is that updates to components from which
cdebootstrap-static is build, needed a manual trigger to be picked up.
By switching to the dynamic variant, that is no longer needed.
We also already had most of the dependencies of the dynamic one already
included in the project.

Signed-off-by: Diederik de Haas <github@cknow.org>