Add signed msgs that prove the archive signing keys.
I already posted the signed messages in issue #64, but they should be
part of the git repo. Fashionably late, but better late then never.
The signing key for raspbian.org was provided by 'plugwash'.
The signing key for raspberrypi.org was provided by 'shiftplusone'.
The KeyIDs are used in update.sh for verifying downloaded packages.

Signed-off-by: Diederik de Haas <github@cknow.org>