Don't DL and check the archive key every time, use the key.
The reasons we/I brought up as a 'reason' to download the key each time
and then verify it, turned out to be unfounded, so stop doing that.
On top of that, we/I didn't implement any check besides the fingerprint,
so all the theoretical threats wouldn't be mitigated anyway.
So essentially @jturner314, @mologie were right all along :)

This also simplifies the code as was also the intend of PR #223, but
now it's even simpler and the gpg key is in binary format, not ascii
armored format. One of the maintainers of APT, 'jak', indicated that
that was prefered. Better to follow the advice of people who are more
knowledgeable on these kind of subjects ;-)

Signed-off-by: Diederik de Haas <github@cknow.org>