Specify alternative gpu_mem options for 256/512MB models.
The (new?) default seems to be 'gpu_mem=64', but that's a big chunck for
especially 256MB models, so use special config settings for those
models. The installer itself doesn't do anything graphically hard, it
only displays text on a console, so it could be set for all.

https://www.raspberrypi.org/documentation/configuration/config-txt/memory.md
is the source for these configuration settings.

Signed-off-by: Diederik de Haas <github@cknow.org>