Implement workaround for low-mem models.
The installer makes a backup (in memory) of itself to /bootfs, but after
restore to the new FAT32 partition doesn't remove that memory copy.
The 256MB model is so memory constrained that it fails, likely due to
an OOM problem. For those models (only), remove the RAM copy and when
the new boot partition gets mounted (to /rootfs/boot), make a
'mount --bind' to /bootfs, so if any code assumes the data/files to be
there, it'll still find them, so it won't break any code.

This fixes #520.

Signed-off-by: Diederik de Haas <github@cknow.org>